+++ CHANGELOG.md
@@ -7,11 +7,10 @@
 
 ## v0.10.0 - 2025-07-02
 
-
-- document `LPCode` using `Hecke`'s (`small_group`) 
 - **(fix)** The gates `SQRTY`, `CXYZ`, `CZYX` were computing phases incorrectly when acting on `I` stabilizers.
 - **(fix)** Paulis with imaginary phases had their phases incorrectly tracked.
 - **(fix)** `rowdecompose` was not accounting for the phase of the input Pauli string, leading to potential errors in non-Clifford functionality.
+- add documentation of using `LPCode` via `Hecke`'s (`small_group`) 
 - Various stabilizer data structures can now be constructed out of sequences of Paulis.
 - Most canonicalization routines can now act on sequences of Paulis.
 - `project*!` family of functions now work on more tableau types.