+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 # News
 
-## v0.10.0 - 2025-06-25
+## v0.10.0 - 2025-07-02
+
 
 - **(fix)** The `parity_checks`, `code_n`, and `code_s` methods were throwing method errors and `code_k` was not defined for the classical `LiftedCode`s.
 - **(fix)** The gates `SQRTY`, `CXYZ`, `CZYX` were computing phases incorrectly when acting on `I` stabilizers.
@@ -13,6 +14,7 @@
 - **(fix)** `rowdecompose` was not accounting for the phase of the input Pauli string, leading to potential errors in non-Clifford functionality.
 - Various stabilizer data structures can now be constructed out of sequences of Paulis.
 - Most canonicalization routines can now act on sequences of Paulis.
+- `project*!` family of functions now work on more tableau types.
 - `QuantumCliffordJuMPExt` is a new extension depending on the JuMP solver suite. It enables computing minimum distance for quantum LDPC codes via Mixed Integer Programming (MIP) using `JuMP`.
 - `QuantumCliffordOscarExt` is a new extension depending on the Oscar CAS. It provides more convenient ways to construct two-block-group-algebra codes and more.
 - Improvements to `GeneralizedStabilizer` API including support for products and tensor product for many Clifford and non-Clifford operators and states.